requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.2
boto3==1.34.0
botocore==1.34.162
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
crispy-bootstrap5==0.7
cryptography==41.0.7
cycler==0.12.1
Django==4.2.7
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
et_xmlfile==2.0.0
filelock==3.20.0
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.7
midas==0.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numpy==1.23.5
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==25.0
pandas==1.5.3
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==3.20.3
psycopg2-binary==2.9.7
pycparser==2.23
PyJWT==2.8.0
pymodbus==3.7.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.3
redis==5.0.1
reportlab==4.0.4
requests==2.32.5
s3transfer==0.9.0
safetensors==0.6.2
schedule==1.2.0
scikit-image==0.21.0
scipy==1.11.4
six==1.17.0
sounddevice==0.5.2
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.5.10
timm==1.0.20
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
